feat: refactor go-template output for the `get` command
This way, it's not necessary to create a slice of `interface{}`
objects, anymore in each `get` sub-command and helps us get rid of
lots of boilerplate code.

This changes the user experience in such a way that the template would
have to iterate over a slice now. So

```
gitlab -p p1 get pipelines -a -o go-template='{{.URL}}'
```
becomes
```
gitlab -p p1 get pipelines -a -o go-template='{{range .}}{{.URL}}{{"\n"}}{{end}}'
```
to achieve the same output.