feat: add `--pre` flag to `update` command (#65)
With this flag, users can now choose to upgrade to the next
pre-release version if there is one available.

Example:

You're running gitlab 3.7.2 at the moment and 3.8.0-beta.1 is released
which is a pre-release version. Running `gitlab update` will not
update to that beta version by default. If you run `gitlab update
--pre`, however, you will get that version installed.